# api/gateway/src/main/resources/static/member-openapi3.yaml
openapi: 3.0.1
info:
  title: 우리집 레시피 API Documentation
  description: Spring REST Docs with SwaggerUI
  version: 0.0.1
servers:
- url: http://localhost:8080/v1
tags: []
paths:
  /member/login:
    post:
      tags:
      - 회원 API
      summary: 로그인
      description: 로그인
      operationId: 로그인 API
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/member-login-930702395'
            examples:
              로그인 API:
                value: "{\"email\":\"admin@gmail.com\",\"password\":\"Admin1234!@\"\
                  }"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/member-token-refresh-1584530892'
              examples:
                로그인 API:
                  value: "{\"code\":200,\"message\":\"정상적으로 처리되었습니다.\",\"data\":{\"\
                    accessToken\":\"eyJhbGciOiJIUzUxMiJ9.eyJpZCI6MSwic3ViIjoiYWRtaW5AZ21haWwuY29tIiwiaWF0IjoxNzI3OTExMjYzLCJleHAiOjE3Mjc5MTMwNjMsInJvbGVzIjoiUk9MRV9BRE1JTiJ9.iMBNqtogso84q-4mjbuTGHVf2L15Czosj1JxaKf-DBZGKS1TUudAzlfmOO4LGxABAnMFEb4TJ7iX6Ixg1pm-FQ\"\
                    ,\"refreshToken\":\"eyJhbGciOiJIUzUxMiJ9.eyJpZCI6MSwic3ViIjoiYWRtaW5AZ21haWwuY29tIiwiaWF0IjoxNzI3OTExMjYzLCJleHAiOjE3Mjc5OTc2NjN9.rFCRYL9tri6bRQqnwyNKRNmsM-AjofK9YyCOizLTEs_qJxFPXZVncWRCrMDnzcpL_NyrFsNXmP6Xx0IL3XZUSQ\"\
                    }}"
  /member/logout:
    post:
      tags:
      - 회원 API
      summary: 로그아웃
      description: 로그아웃
      operationId: 로그아웃 API
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/member-email-auth-request1575598845'
              examples:
                로그아웃 API:
                  value: "{\"code\":200,\"message\":\"정상적으로 로그아웃 처리되었습니다.\"}"
      security:
      - bearerAuthJWT: []
  /member/register:
    post:
      tags:
      - 회원 API
      summary: 회원 가입
      description: 회원 가입
      operationId: 회원 가입 API
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/member-register-1137321134'
            examples:
              회원 가입 API:
                value: "{\"email\":\"test@example.com\",\"password\":\"Pass123!@\"\
                  ,\"passwordConfirm\":\"Pass123!@\",\"nickname\":\"testUser\",\"\
                  phoneNumber\":\"010-1234-5678\",\"name\":\"테스트\",\"introduce\":\"\
                  \"}"
      responses:
        "201":
          description: "201"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/member-register1233141654'
              examples:
                회원 가입 API:
                  value: "{\"code\":201,\"message\":\"정상적으로 생성되었습니다.\",\"data\":{\"\
                    id\":4}}"
  /member/follow/{followingId}:
    post:
      tags:
      - 회원 API
      summary: 회원 팔로우
      description: 회원 팔로우
      operationId: 회원 팔로우 API
      parameters:
      - name: followingId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/member-email-auth-request1575598845'
              examples:
                회원 팔로우 API:
                  value: "{\"code\":200,\"message\":\"정상적으로 처리되었습니다.\"}"
      security:
      - bearerAuthJWT: []
  /member/me/profile:
    get:
      tags:
      - 회원 API
      summary: 내 프로필 조회
      description: 내 프로필 조회
      operationId: 내 프로필 조회 API
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/member-me-profile-1720120988'
              examples:
                내 프로필 조회 API:
                  value: "{\"code\":200,\"message\":\"정상적으로 처리되었습니다.\",\"data\":{\"\
                    id\":1,\"email\":\"admin@gmail.com\",\"nickname\":\"ê´€ë¦¬ìž\x90\
                    \ ë‹‰ë„¤ìž„\",\"phoneNumber\":\"010-000-0000\",\"name\":\"ê´€ë\
                    ¦¬ìž\x90\",\"profileImage\":\"\",\"introduce\":null}}"
      security:
      - bearerAuthJWT: []
    post:
      tags:
      - 회원 API
      summary: 내 프로필 업데이트
      description: 내 프로필 업데이트
      operationId: 내 프로필 업데이트 API
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/member-me-profile901414125'
              examples:
                내 프로필 업데이트 API:
                  value: "{\"code\":200,\"message\":\"프로필 수정이 정상적으로 처리되었습니다.\",\"\
                    data\":{\"id\":1}}"
      security:
      - bearerAuthJWT: []
  /member/search/{nickname}:
    get:
      tags:
      - 회원 API
      summary: 회원 검색
      description: 회원 검색
      operationId: 회원 검색 API
      parameters:
      - name: nickname
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/member-search-nickname-885510247'
              examples:
                회원 검색 API:
                  value: "{\"code\":200,\"message\":\"정상적으로 처리되었습니다.\",\"data\":{\"\
                    totalPages\":0,\"totalElements\":0,\"first\":true,\"last\":true,\"\
                    sort\":{\"empty\":true,\"unsorted\":true,\"sorted\":false},\"\
                    size\":10,\"content\":[],\"number\":0,\"pageable\":{\"pageNumber\"\
                    :0,\"pageSize\":10,\"sort\":{\"empty\":true,\"unsorted\":true,\"\
                    sorted\":false},\"offset\":0,\"unpaged\":false,\"paged\":true},\"\
                    numberOfElements\":0,\"empty\":true}}"
      security:
      - bearerAuthJWT: []
  /member/token/refresh:
    get:
      tags:
      - 회원 API
      summary: 토큰 재발급
      description: 토큰 재발급
      operationId: 토큰 재발급 API
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/member-token-refresh-1584530892'
              examples:
                토큰 재발급 API:
                  value: "{\"code\":200,\"message\":\"정상적으로 처리되었습니다.\",\"data\":{\"\
                    accessToken\":\"eyJhbGciOiJIUzUxMiJ9.eyJpZCI6MSwic3ViIjoiYWRtaW5AZ21haWwuY29tIiwiaWF0IjoxNzI3OTExMjY0LCJleHAiOjE3Mjc5MTMwNjQsInJvbGVzIjoiIn0.MhvMS8jj0HPtVQGdI7yN5a6a8DFxP5oQixCnyIa97l3Q2THmCkO4d92vS27EfKdBo6bgDrCySLUx3wm6zQNyYA\"\
                    ,\"refreshToken\":\"eyJhbGciOiJIUzUxMiJ9.eyJpZCI6MSwic3ViIjoiYWRtaW5AZ21haWwuY29tIiwiaWF0IjoxNzI3OTExMjY0LCJleHAiOjE3Mjc5OTc2NjR9.jRaoXgmFliatE218aoHXKuOCkRIoD5h_Peo12icplU54fxiMLZ39bqVBeB3NPy3qNHze4xVsQRjGAwh9U3pr0w\"\
                    }}"
      security:
      - bearerAuthJWT: []
  /member/unfollow/{followingId}:
    post:
      tags:
      - 회원 API
      summary: 회원 언팔로우
      description: 회원 언팔로우
      operationId: 회원 언팔로우 API
      parameters:
      - name: followingId
        in: path
        description: ""
        required: true
        schema:
          type: string
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/member-email-auth-request1575598845'
              examples:
                회원 언팔로우 API:
                  value: "{\"code\":200,\"message\":\"정상적으로 처리되었습니다.\"}"
      security:
      - bearerAuthJWT: []
  /member/email/auth/confirm:
    post:
      tags:
      - 회원 API
      summary: 이메일 인증 코드 확인
      description: 이메일 인증 코드 확인
      operationId: 이메일 인증 코드 확인 API
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/member-email-auth-confirm2121084288'
            examples:
              이메일 인증 코드 확인 API:
                value: "{\"email\":\"seongo0521@gmail.com\",\"authCode\":\"123456\"\
                  }"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/member-email-auth-request1575598845'
              examples:
                이메일 인증 코드 확인 API:
                  value: "{\"code\":200,\"message\":\"이메일 인증이 정상적으로 처리되었습니다.\"}"
  /member/email/auth/request:
    post:
      tags:
      - 회원 API
      summary: 이메일 인증 코드 요청
      description: 이메일 인증 코드 요청
      operationId: 이메일 인증 코드 요청 API
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/member-email-auth-request-1523986593'
            examples:
              이메일 인증 코드 요청 API:
                value: "{\"email\":\"seongo0521@gmail.com\"}"
      responses:
        "200":
          description: "200"
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/member-email-auth-request1575598845'
              examples:
                이메일 인증 코드 요청 API:
                  value: "{\"code\":200,\"message\":\"인증 메일이 정상적으로 발송되었습니다.\"}"
components:
  schemas:
    member-search-nickname-885510247:
      type: object
      properties:
        code:
          type: number
          description: 상태 코드
        data:
          type: object
          description: 회원 검색 데이터
        message:
          type: string
          description: 상태 메시지
    member-login-930702395:
      type: object
      properties:
        password:
          type: string
          description: 패스워드
        email:
          type: string
          description: 이메일(아이디)
    member-email-auth-request1575598845:
      type: object
      properties:
        code:
          type: number
          description: 상태 코드
        message:
          type: string
          description: 상태 메시지
    member-email-auth-request-1523986593:
      type: object
      properties:
        email:
          type: string
          description: 이메일(아이디)
    member-register1233141654:
      type: object
      properties:
        code:
          type: number
          description: 상태 코드
        data:
          type: object
          properties:
            id:
              type: number
              description: 아이디 고유번호
        message:
          type: string
          description: 상태 메시지
    member-register-1137321134:
      type: object
      properties:
        password:
          type: string
          description: 패스워드
        phoneNumber:
          type: string
          description: 핸드폰 번호
        passwordConfirm:
          type: string
          description: 패스워드 확인
        introduce:
          type: string
          description: 소개
        name:
          type: string
          description: 이름
        nickname:
          type: string
          description: 닉네임
        email:
          type: string
          description: 이메일(아이디)
    member-me-profile901414125:
      type: object
      properties:
        code:
          type: number
          description: 상태 코드
        data:
          type: object
          properties:
            id:
              type: number
              description: 업데이트된 회원의 ID
          description: 업데이트된 프로필 데이터
        message:
          type: string
          description: 상태 메시지
    member-token-refresh-1584530892:
      type: object
      properties:
        code:
          type: number
          description: 상태 코드
        data:
          type: object
          description: 토큰 데이터
        message:
          type: string
          description: 상태 메시지
    member-me-profile-1720120988:
      type: object
      properties:
        code:
          type: number
          description: 상태 코드
        data:
          type: object
          description: 내 프로필 데이터
        message:
          type: string
          description: 상태 메시지
    member-email-auth-confirm2121084288:
      type: object
      properties:
        authCode:
          type: string
          description: 인증 코드
        email:
          type: string
          description: 이메일(아이디)
  securitySchemes:
    bearerAuthJWT:
      type: http
      scheme: bearer
      bearerFormat: JWT